# MainApp/docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
  models.Dialog:
    properties:
      dialog_id:
        readOnly: true
        type: integer
      messages:
        items:
          $ref: '#/definitions/models.Message'
        readOnly: true
        type: array
      userId1:
        type: integer
      userId2:
        type: integer
    type: object
...
  models.Message:
    properties:
      body:
        type: string
      created_at:
        type: string
      dialog_id:
        type: integer
      id:
        readOnly: true
        type: integer
      receiver_id:
        type: integer
      sender_id:
        readOnly: true
        type: integer
    type: object
...
  pkg.Response:
    properties:
      body: {}
    type: object
host: 89.208.197.127:8080
info:
  contact: {}
  title: WS Swagger API
  version: "1.0"
paths:
...
  /chat:
    get:
      description: get all dialogs
      produces:
      - application/json
      responses:
        "200":
          description: success get dialogs
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/models.Dialog'
                  type: array
              type: object
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: GetAllDialogs
      tags:
      - chat
  /chat/{id}:
    get:
      description: get dialog
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get dialog
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.Dialog'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find chat with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: GetDialog
      tags:
      - chat
  /chat/send_message:
    post:
      consumes:
      - application/json
      description: send message
      parameters:
      - description: message data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Message'
      produces:
      - application/json
      responses:
        "200":
          description: success send message
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.Message'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find item with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: SendMessage
      tags:
      - chat
  /chat/user/{id}:
    get:
      description: get dialog
      parameters:
      - description: Friend ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get dialog
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/models.Dialog'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find chat with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: GetDialogByUsers
      tags:
      - chat
...
swagger: "2.0"